Render discount prices for artifacts

+++ src/main/java/com/codecool/quest_store/dao/ItemDao.java
@@ -9,6 +9,6 @@ public interface ItemDao extends Dao<Item> {
     List<Item> getAllItems() throws DaoException;
     Item getItemById(int itemId) throws DaoException;
 
-    float getDiscount() throws DaoException;
+    int getDiscount() throws DaoException;
 }
